services/debate_moderator/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
websockets>=12.0,<14.0
anthropic>=0.42.0,<1.0
httpx>=0.28.0,<1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
deepgram-sdk==3.8.0
numpy==1.26.4
